+++ src/main/java/se/lingonskogen/em2012/domain/Prediction.java
@@ -19,9 +19,9 @@ public class Prediction extends Bean {
 
 	private String gameId;
 
-	private Integer homeScore;
+	private Long homeScore;
 
-	private Integer awayScore;
+	private Long awayScore;
 
 	public String getGroupId() {
 		return groupId;
@@ -63,19 +63,19 @@ public class Prediction extends Bean {
 		this.gameId = gameId;
 	}
 
-	public Integer getHomeScore() {
+	public Long getHomeScore() {
 		return homeScore;
 	}
 
-	public void setHomeScore(Integer homeScore) {
+	public void setHomeScore(Long homeScore) {
 		this.homeScore = homeScore;
 	}
 
-	public Integer getAwayScore() {
+	public Long getAwayScore() {
 		return awayScore;
 	}
 
-	public void setAwayScore(Integer awayScore) {
+	public void setAwayScore(Long awayScore) {
 		this.awayScore = awayScore;
 	}
