+++ src/main/java/se/lingonskogen/em2012/services/PredictionService.java
@@ -8,7 +8,7 @@ import se.lingonskogen.em2012.domain.Prediction;
 public interface PredictionService {
 	String createPrediction(final Prediction prediction) throws DaoException;
 	Prediction newInstance(final String groupId, final String userId, final String couponId, 
-			final String tournamentId, final String gameId, final Integer homeScore, final Integer awayScore);
+			final String tournamentId, final String gameId, final Long homeScore, final Long awayScore);
 	List<Prediction> getAvailablePredictions();
 
 	List<Prediction> findPredictions(final String groupId, final String userId, final String couponId);
